refactor: Move Exception Util File at Domain Module
