+++ domain/build.gradle.kts
@@ -38,6 +38,7 @@ dependencies {
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
+    implementation(libs.firebase.firestore)
     implementation(libs.kotlinx.serialization.json)
 
     testImplementation(libs.junit)
