+++ domain/src/main/java/com/kitching/domain/exception/ErrorConst.kt
@@ -1,4 +1,4 @@
-package com.kitching.data.util
+package com.kitching.domain.exception
 
 const val NETWORK_EXCEPTION_MESSAGE = "네트워크 오류가 발생했습니다.  잠시 후 다시 시도해주세요."
 const val NETWORK_ERROR_CODE = "NETWORK_ERROR"
